rename add-tags-input component to add-tags
